+++ backend/src/main/java/com/prograngers/backend/Encrypt.java
@@ -1,5 +1,7 @@
 package com.prograngers.backend;
 
+import com.prograngers.backend.dto.FailToDecodeException;
+import com.prograngers.backend.dto.FailToEncodeException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
